+++ tactics2d/sensor/__init__.py
@@ -2,12 +2,22 @@
 # Copyright (C) 2024, Tactics2D Authors. Released under the GNU GPLv3.
 # @File: __init__.py
 # @Description: Initialize the sensor module.
-# @Author: Yueyuan Li
+# @Author: Tactics2D Team
 # @Version: 1.0.0
 
+# isort: off
 from .camera import TopDownCamera
+from .sensor_base import SensorBase
+from .carla_sensor_base import CarlaSensorBase
 from .lidar import SingleLineLidar
 from .render_manager import RenderManager
-from .sensor_base import SensorBase
 
-__all__ = ["SensorBase", "TopDownCamera", "SingleLineLidar", "RenderManager"]
+__all__ = [
+    "SensorBase",
+    "TopDownCamera",
+    "SingleLineLidar",
+    "RenderManager",
+    "SensorBase",
+    "CarlaSensorBase",
+]
+# isort: on